Implemented CRUD tags

+++ algonetAPI/src/main/java/com/algonet/algonetapi/repositories/ProblemRepository.java
@@ -3,6 +3,6 @@ package com.algonet.algonetapi.repositories;
 import com.algonet.algonetapi.Models.entities.Problem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProblemRepository extends JpaRepository<Problem, Long> {
+public interface ProblemRepository extends JpaRepository<Problem, Integer> {
 
 }
