+++ algonetAPI/src/main/java/com/algonet/algonetapi/repositories/TestRepository.java
@@ -3,5 +3,5 @@ package com.algonet.algonetapi.repositories;
 import com.algonet.algonetapi.Models.entities.Test;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestRepository extends JpaRepository<Test, Long> {
+public interface TestRepository extends JpaRepository<Test, Integer> {
 }
